Adding Gyro2 and Gyro3 to the scope of Gyro concern
Adding tests;
Fixing problem with blocknumber from old tests;
