+++ balancer-js/src/modules/pools/pool-types/concerns/gyro/spotPrice.concern.ts
@@ -1,7 +1,7 @@
 import { SpotPriceConcern } from '@/modules/pools/pool-types/concerns/types';
 import { Pool } from '@/types';
 
-export class GyroESpotPriceConcern implements SpotPriceConcern {
+export class GyroSpotPriceConcern implements SpotPriceConcern {
   calcPoolSpotPrice(tokenIn: string, tokenOut: string, pool: Pool): string {
     console.log(tokenIn, tokenOut, pool);
     throw new Error('Not implemented');
